Refactor VLM modules for internvl-llava (#2797)
